feat: make rehoboam a native node package
grpc is nice but having to start a separate service is annoying and it makes deployment harder. now rehoboam is just a node package consumed by sotast - pre/post processing model inputs/outputs is done in rust so its fast, but everything else like downloading models and unloading after a delay can be done from the typescript side where its a little easier and can integrate more closely with other features.

this does lose the one upside of grpc though, which is running the rehoboam service on a more powerful remote machine (like your gaming pc) and running the sotast+blahaj services on a less powerful machine. but i might come back to that in the future.