feat: use tiny webps for previews
this was partially done in a previous commit but it was not finished so this is the other half. its blurred server-side which has some artifacts on the client but makes it perform much better, and is much more versatile because its just a normal image. this seems to fix almost all the issues with the previous version, including huge SSR HTML payloads.